* commit 'f2a5ae0411ddd0638897e9ebb786a0947f9a7f97':
  更细嫩算法
  no message
  接入glTF解析，编译通过。

+++ SVEngine/src/core/SVImageUsingMove.h
@@ -26,7 +26,7 @@ namespace sv {
             
             SVArray<FVec2> m_targetArray;
             
-            SVArray<f32>  m_weightArray;
+            SVArray<f64>  m_weightArray;
             
             FVec2 MLS(const FVec2& t);
